+++ templates/page/FeaturePage.xaml.cs
@@ -1,0 +1,10 @@
+namespace MauiApp1;
+
+public partial class FeaturePage : ContentPage
+{
+    public FeaturePage(FeatureViewModel viewModel)
+    {
+        InitializeComponent();
+        BindingContext = viewModel;
+    }
+}